Migration to latexmk
latexmk is able to properly track changes on all the sources used in a
LaTeX project, and it is also able to detect whether to run again latex
or not. In this way, model-doc-generator.pl does not have to blindly
run latex, having faith on a high number of re-runs.